+++ packages/graph/hash_graph/lib/graph/src/store/postgres/ontology/property_type/mod.rs
@@ -36,7 +36,8 @@ impl<C: AsClient> PropertyTypeStore for PostgresStore<C> {
             .insert_property_type_references(&property_type, version_id)
             .await
             .change_context(InsertionError)
-            .attach_printable("Could not insert references for property type")
+            // TODO THIS COMMENT SHOULDN'T MAKE IT INTO A PR. We need to update other errors to have similar info 
+            .attach_printable_lazy(|| format!("Could not insert references for property type: {}", property_type.id()))
             .attach_lazy(|| property_type.clone())?;
 
         transaction
